Modified folders structure & Ran prettier
